[MDETR] Move model components into models/mdetr
Summary: Mostly just moving around existing components. Create a new directory models/mdetr
and put text and image encoders in there. Also move transformer components into their own file.
This is prep for the next PR, which will move task-specific models into mdetr/model.py, so taking
the transformer stuff out will make the file clearer.

Test plan:

```
python -m pytest -vv

```

[ghstack-poisoned]
