+++ test/models/mdetr/test_transformer.py
@@ -8,10 +8,10 @@ import pytest
 import torch
 
 from test.test_utils import assert_expected, set_rng_seed
-from torchmultimodal.models.mdetr import mdetr_resnet101, MDETRTransformer
+from torchmultimodal.models.mdetr.transformer import MDETRTransformer
 
 
-class TestMDETR:
+class TestMDETRTransformer:
     @pytest.fixture(autouse=True)
     def rng(self):
         set_rng_seed(0)
@@ -25,10 +25,6 @@ class TestMDETR:
         return 4
 
     @pytest.fixture(autouse=True)
-    def text_dim(self):
-        return 5
-
-    @pytest.fixture(autouse=True)
     def mm_dim(self):
         return 11
 
@@ -39,14 +35,6 @@ class TestMDETR:
     @pytest.fixture(autouse=True)
     def num_decoder_layers(self):
         return 6
-
-    @pytest.fixture(autouse=True)
-    def num_queries_full(self):
-        return 100
-
-    @pytest.fixture(autouse=True)
-    def num_classes_full(self):
-        return 255
 
     @pytest.fixture()
     def src(self, mm_dim, batch_size, embedding_dim):
@@ -69,14 +57,6 @@ class TestMDETR:
         return torch.randn(mm_dim, batch_size, embedding_dim)
 
     @pytest.fixture()
-    def text_memory(self, text_dim, batch_size, embedding_dim):
-        return torch.randn(text_dim, batch_size, embedding_dim)
-
-    @pytest.fixture()
-    def text_memory_key_padding_mask(self, batch_size, text_dim):
-        return torch.randint(0, 2, (batch_size, text_dim))
-
-    @pytest.fixture()
     def memory_key_padding_mask(self, batch_size, mm_dim):
         return torch.randint(0, 2, (batch_size, mm_dim))
 
@@ -85,50 +65,12 @@ class TestMDETR:
         return torch.randn(num_queries, batch_size, embedding_dim)
 
     @pytest.fixture()
-    def test_tensors(self):
-        return torch.rand(2, 3, 64, 64).unbind(dim=0)
-
-    @pytest.fixture()
-    def input_ids(self):
-        return torch.Tensor(
-            [
-                [0, 100, 64, 192, 5, 3778, 2, 1, 1, 1, 1, 1, 1, 1, 1, 1],
-                [
-                    0,
-                    1708,
-                    190,
-                    114,
-                    38,
-                    1395,
-                    192,
-                    5,
-                    3778,
-                    6,
-                    38,
-                    216,
-                    14,
-                    24,
-                    8785,
-                    2,
-                ],
-            ]
-        ).to(dtype=torch.long)
-
-    @pytest.fixture()
     def transformer(self, embedding_dim, num_decoder_layers):
         transformer = MDETRTransformer(
             d_model=embedding_dim, num_decoder_layers=num_decoder_layers
         )
         transformer.eval()
         return transformer
-
-    @pytest.fixture()
-    def mdetr(self, num_queries_full, num_classes_full):
-        mdetr = mdetr_resnet101(
-            num_queries=num_queries_full, num_classes=num_classes_full
-        )
-        mdetr.eval()
-        return mdetr
 
     def test_transformer_encoder(
         self,
@@ -177,52 +119,3 @@ class TestMDETR:
             [[-1.6592, 0.3761], [-2.5531, 0.7192], [-1.2693, 0.3763], [-1.1510, 0.1224]]
         )
         assert_expected(actual[1, :, :, 1], expected, rtol=0, atol=1e-3)
-
-    def test_full_mdetr_model(
-        self,
-        mdetr,
-        test_tensors,
-        input_ids,
-        batch_size,
-        num_queries_full,
-        num_classes_full,
-    ):
-        out = mdetr(test_tensors, input_ids)
-        logits_actual = out.pred_logits
-        boxes_actual = out.pred_boxes
-        logits_expected = torch.Tensor(
-            [
-                -0.0145,
-                0.0121,
-                0.0270,
-                0.0310,
-                0.0072,
-                -0.0002,
-                0.0100,
-                0.0012,
-                0.0290,
-                0.0067,
-            ]
-        )
-        boxes_expected = torch.Tensor(
-            [
-                0.4896,
-                0.4898,
-                0.4897,
-                0.4900,
-                0.4894,
-                0.4895,
-                0.4897,
-                0.4908,
-                0.4902,
-                0.4899,
-            ]
-        )
-        assert logits_actual.size() == (
-            batch_size,
-            num_queries_full,
-            num_classes_full + 1,
-        )
-        assert boxes_actual.size() == (batch_size, num_queries_full, 4)
-        assert_expected(logits_actual[1, :10, 1], logits_expected, rtol=0, atol=1e-3)
-        assert_expected(boxes_actual[1, :10, 1], boxes_expected, rtol=0, atol=1e-3)
