+++ torchmultimodal/models/mdetr/image_encoder.py
